requirements.txt
streamlit>=1.28.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
chromadb>=0.4.0
pandas>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
pydantic>=2.0.0
